aws_free_tier/3_app_runner/paris_cafe/cafes/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import CafeteriaViewSet

router = DefaultRouter()
router.register(r'cafes', CafeteriaViewSet)

urlpatterns = [
    path('', include(router.urls)),
]