aws_free_tier/3_app_runner/paris_cafe/requirements.txt
Django
djangorestframework
django-cors-headers
django-filter
markdown